Moved routecontroller to lib
